+++ src/ble/alerts/index.ts
@@ -4,7 +4,8 @@
 import { bluetoothNotAvailable } from './BluetoothNotAvailable';
 import { missingService } from './MissingService';
 import { noGatt } from './NoGatt';
+import { noHub } from './NoHub';
 import { noWebBluetooth } from './NoWebBluetooth';
 
 // gathers all of the alert creation functions for passing up to the top level
-export default { bluetoothNotAvailable, missingService, noGatt, noWebBluetooth };
+export default { bluetoothNotAvailable, missingService, noGatt, noHub, noWebBluetooth };