+++ src/firmware/installPybricksDialog/actions.ts
@@ -1,0 +1,24 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2022 The Pybricks Authors
+
+import { createAction } from '../../actions';
+
+/** Actions that request the install Pybricks firmware dialog to be shown. */
+export const firmwareInstallPybricksDialogShow = createAction(() => ({
+    type: 'firmware.installPybricksDialog.action.show',
+}));
+
+/** Actions that indicates the user accepted the install Pybricks firmware dialog. */
+export const firmwareInstallPybricksDialogAccept = createAction(
+    (firmwareZip: ArrayBuffer, customProgram: string | undefined, hubName: string) => ({
+        type: 'firmware.installPybricksDialog.action.accept',
+        firmwareZip,
+        customProgram,
+        hubName,
+    }),
+);
+
+/** Actions that indicates the user canceled the install Pybricks firmware dialog. */
+export const firmwareInstallPybricksDialogCancel = createAction(() => ({
+    type: 'firmware.installPybricksDialog.action.cancel',
+}));