+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - Added better error message when no files to backup ([support#681]).
+- Added multi-step firmware flashing dialog.
 
 ### Fixed
 - Fixed deleting files that are not open in the editor.